Smart Home
This project is a smart home system. It has been developed using Arduino and the circuit design is done using Proteus.
Operation:

When the flame sensor is activated, the buzzer starts to sound.
If the PIR sensor detects motion, the lamp turns on.
If the correct password is entered on the keypad, the green LED lights up. Otherwise, the red LED lights up.
The temperature value measured by the temperature sensor is displayed on the LCD screen.